Get UI essentially functional
The calculator works! You can do all the basic maths with it!

There are a few small weird glitchy things, some design decisions to be
made, and possible additional features. It should also be tested more
extensively to make sure I accounted for various use instances.

I also need to figure out how to batch process event listeners and have
it actually work because all these individual ones are making the code
twice as long as it could be and probably slowing it down a bunch.