Allow longer numbers and update color scheme
User can enter numbers up to 15 digits. Text flows off screen to the
right as you enter really long numbers, and the solution displays
without the partial/overflowed number to the right of it. This seems
like a good compromise for the limits of a simple calculator: you can
see that your entries are still entering as you type, and when you hit
"equals" you get the solution nice and clean all by itself, and you can
use it to start the next calculation. If you stick with smaller numbers
that produce a smaller solution, you still get to see the whole
calculation on the display.

Very large or very small solutions now display with the exponent
signifier "e" and fit on the screen. You cannot use those to start a new
calculation like you can with smaller numbers, but I think that's fair.

Code is also cleaned up immensely, event listeners streamlined.

Color scheme is now turquoise / midnight blue / orange-red / olive.
Might be jewel tone overkill, but when is it not with me? If not a
calculator, this would be a fantastic color scheme for a striped sock.